+++ ChangeLog.md
@@ -31,6 +31,7 @@
 * Support pour les personnages invisibles en combat
 * Meilleur support de la Prouesse du guerrier
 * Meilleur support de la Parade de projectiles du moine
+* Meilleur support du Tour de force
 
 ### Corrections de bugs
 * Calcul de score d'attaque corrigé pour les Terres d'Arran.